提取接口
